got recipe create to work

+++ app/recipes/actions.ts
@@ -4,12 +4,14 @@ import { prisma } from "@/lib/client";
 import { redirect } from "next/navigation";
 import { zfd } from "zod-form-data";
 import { z } from "zod";
+import slugify from "slugify";
 
 const schema = zfd.formData({
-  id: zfd.numeric(),
+  id: zfd.numeric(z.number().optional()),
   name: zfd.text(),
   description: zfd.text(z.string().optional()),
-  styleIdentifer: zfd.text(z.string().optional()),
+  authorEmail: zfd.text(),
+  styleIdentifer: zfd.text(),
   fermentables: z
     .object({
       //recipeId: zfd.numeric(z.number()),
@@ -17,8 +19,8 @@ const schema = zfd.formData({
       amount: zfd.numeric(z.number().min(0)),
       amountType: z.nativeEnum(MassUnit).default(MassUnit.oz),
     })
-    .array(),
-
+    .array()
+    .optional(),
   hops: z
     .object({
       //recipeId: zfd.numeric(z.number()),
@@ -29,25 +31,45 @@ const schema = zfd.formData({
       durationType: z.nativeEnum(TimeUnit).default(TimeUnit.min),
       //amountType: z.enum(["kg", "g", "oz", "lb"]).default("kg"),
     })
-    .array(),
+    .array()
+    .optional(),
 });
 export async function updateRecipe(formData: FormData) {
-  const data = schema.parse(formData);
-  const res = await prisma.recipe.update({
-    where: {
-      id: data.id,
-    },
-    data: {
-      ...data,
-      fermentables: {
-        deleteMany: {},
-        createMany: { data: data.fermentables },
+  const { id, authorEmail, styleIdentifer, ...data } = schema.parse(formData);
+  let res;
+  if (id) {
+    res = await prisma.recipe.update({
+      where: {
+        id,
       },
-      hops: {
-        deleteMany: {},
-        createMany: { data: data.hops },
+      data: {
+        ...data,
+        fermentables: {
+          deleteMany: {},
+          createMany: { data: data.fermentables || [] },
+        },
+        hops: {
+          deleteMany: {},
+          createMany: { data: data.hops || [] },
+        },
       },
-    },
-  });
+    });
+  } else {
+    res = await prisma.recipe.create({
+      data: {
+        ...data,
+        style: { connect: { identifier: styleIdentifer } },
+        author: {
+          connect: {
+            email: authorEmail,
+          },
+        },
+
+        slug: slugify(data.name, { lower: true }),
+        fermentables: { createMany: { data: data.fermentables || [] } },
+        hops: { createMany: { data: data.hops || [] } },
+      },
+    });
+  }
   redirect(`/recipes/${res.id}`);
 }
